Module File_impl.Iter_block_blit

Abstract model of blitting from blks to buffer.

Spec: a seq of blits equiv to a single blit, but at most 1 blit is not block aligned, and at most 2 blits are not blk sized; also, minimal number of blits (no overlapping blits).

This version assumes that the file is backed by a map from blk_index -> blk_id (although for the test code, we ignore blk_id and just work directly with bytes.)

read_blk: given a block index, we look up the actual blk_id in the map, and read that block; return empty blk if no blk; ignore size

alloc_and_write_blk: for the situation where we definitely want to allocate a new blk; we have a blk_index, a blk, and we allocate a blk_id, write the blk, insert the (index,blk_id) into the map, then return the blk_id

rewrite_blk_or_alloc: given a blk_index and a blk_id (of the previous version of the blk) and the new blk, we attempt to rewrite that blk directly; if for some reason we cannot mutate the old blk, we allocate a new blk_id, write the new blk, insert the (index,blk_id) into the map

truncate n: drop the blk-index map entries for all blocks after byte index n

For the usage with file_impl, we assume the inode is locked (so we can update the inode using set_state).

type ('buf, 't) ops = {

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> ('buf, 't) Tjr_monad.m;

}

val make : monad_ops:'t Tjr_monad.monad_ops -> buf_ops:'buf Tjr_fs_shared.Buffers_from_btree.buf_ops -> blk_ops:'blk Tjr_fs_shared.blk_ops -> int_index_iso:(int, 'i) iso -> read_blk:('i Tjr_fs_shared.Int_like.index_ -> ('blk_id * 'blk, 't) Tjr_monad.m) -> alloc_and_write_blk:('i Tjr_fs_shared.Int_like.index_ -> 'blk -> (unit, 't) Tjr_monad.m) -> rewrite_blk_or_alloc:('i Tjr_fs_shared.Int_like.index_ -> ('blk_id * 'blk) -> (unit, 't) Tjr_monad.m) -> ('buf, 't) ops

NOTE we want make to be (as far as possible) independent of the updates to the blk-index map. NOTE that blk_id is essentially completely abstract here, and only used when rewriting.

val test : unit -> unit