Module Tjr_mini_impfs__File_impl

A simple file implementation. Primarily this is here so that we can use it to marshall the free list

Terminology: (file) block-index map: the map from blk index to blk_id

We implement a file as a map from blk_index to blk. The map is typically implemented by a B-tree or similar.

The inode contains the file size and the root of the blk_index map.

New blocks are allocated as needed (by default the map is empty, and a block is only allocated when that block is written to; so files are "sparse").

The "iterated block read/write" component has been abstracted out because it appears in a lot of other places.

The components we require are as follows (see "make" function below; NOTE missing free, since this is not currently implemented for file.truncate).

let make (type fid blk blk_id t) 
      ~monad_ops
      ~(blk_ops       : blk blk_ops)
      ~(blk_dev_ops   : (blk_id,blk,t) blk_dev_ops)
      ~(blk_index_map : (int,blk_id,blk_id,t)pre_map_ops)
      ~(with_inode    : ((fid,blk_id)inode,t)with_state)
      ~(alloc         : unit -> (blk_id,t)m)
  =

TODO:

b-tree to implement delete_after (or some alternative approach)

validate against "normal" file semantics

ensure the tests are run by some top-level test executable

implement freeing of blocks after truncate

maybe worth distinguishing the blk_id used by the map from the blk_id used to store data

type 'fid file_id = {

fid : 'fid;

}

module Inode : sig ... end

module Pre_map_ops : sig ... end

Pre_map_ops: because the file root is stored in the inode, we need to access the B-tree using explicit root passing (and then separately update the inode); this is to ensure that the inode update is atomic (via with_inode). We could store the root in a separate block but this is a bit too inefficient.

type pread_error =

| Pread_error of string

type pwrite_error =

| Pwrite_error of string

type ('buf, 't) file_ops = {

size : unit -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> ((Tjr_fs_shared.Int_like.size, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

truncate : size:Tjr_fs_shared.Int_like.size -> (unit, 't) Tjr_monad.m;

}

Standard file operations, pwrite, pread, size and truncate.

NOTE we expect buf to be string for the functional version; for mutable buffers we may want to pass the buffer in as a parameter?

FIXME for pwrite, we always return src_len since all bytes are written (unless there is an error of course). So perhaps return unit.

For pread, we always return a buffer of length len.

type ('a, 'b) iso = {

a_to_b : 'a -> 'b;

b_to_a : 'b -> 'a;

}

val pread_check : size:int Tjr_fs_shared.Int_like.size_ -> off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'buf Tjr_fs_shared.Buffers_from_btree.buf_ops -> src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (unit, string) Stdlib.result

Check arguments to pwrite

module Iter_block_blit : sig ... end

Abstract model of blitting from blks to buffer.

val make : monad_ops:'t Tjr_monad.monad_ops -> blk_ops:'blk Tjr_fs_shared.blk_ops -> blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_index_map:(int, 'blk_id, 'blk_id, 't) Pre_map_ops.pre_map_ops -> with_inode:(('fid, 'blk_id) Inode.inode, 't) Tjr_monad.with_state -> alloc:(unit -> ('blk_id, 't) Tjr_monad.m) -> (Tjr_fs_shared.Buffers_from_btree.Safe_bytes_buf.bytes_buf, 't) file_ops

NOTE The pread,pwrite functions that result will lock the inode whilst executing, in order to udpate the B-tree root

let make (type fid blk blk_id t) 
      ~monad_ops
      ~(blk_ops       : blk blk_ops)
      ~(blk_dev_ops   : (blk_id,blk,t) blk_dev_ops)
      ~(blk_index_map : (int,blk_id,blk_id,t)pre_map_ops)
      ~(with_inode    : ((fid,blk_id)inode,t)with_state)
      ~(alloc         : unit -> (blk_id,t)m)
  =

val test : unit -> unit

Test with

imperative monad

blk_sz 2

blks are strings

blk_dev is in-memory, map from blk_id to blk

blk_id is blk_id_as_int

blk_index map is a ref (to a map from int to blk_id)

blk_index_map_root is a dummy blk_id -1